Fixed name conflict
The TimerClass had a local `SpeedFaktor` variable, which is used by the sync function.
This leads to a desync between the local and the global variable and can cause the game
speed to be calculated wrong.

Example:
The game speeds up when it runs with a low framerate (the speed factor is changed
to have more accurate calculations, but the sync function still uses the "old" variable)
This can be easily repuduced by starting the game in debug mode and typing
```
maxfps 10
```
in the console to limit the framerate.